test(moe): consolidate and trim shared-expert coverage
Folds tests/moe/test_unified_moe_cache_key.py into test_unified_moe.py --
these are unified-runner contract tests, same as the runner-support and
shared-expert gating classes already there -- and cuts redundant
parametrization.

The cache-key tests fanned out over five runners, but all of them execute
the same base _cache_key_extras(); no backend overrides
_backend_cache_key_parts, so four re-tested identical code. Two remain, one
with two quant variants and one with a single variant.
test_every_registered_runner_defines_stable_extras still covers what the
fan-out was guarding: a backend that does not inherit the contract. The
trace dispatch test likewise dropped from five non-DeepSeekV3 routing
methods to two, since they take the same branch.

105 tests to 75, with no loss of coverage: removing intermediate_size from
the cache key still fails the reduced set.

Also simplifies the get_cache_key_extras() docstring.
