+++ flashinfer/autotuner/autotuner.py
@@ -572,15 +572,11 @@ class TunableRunner(ABC):
         on dtype, is-None flags, or scalar-argument values -- not on
         per-tensor content).
 
-        Every element must also have a **stable ``str()`` across processes**.
-        Hashability alone is not sufficient: ``ProfilingCacheKey.file_key``
-        renders this tuple with ``str()`` to key the on-disk cache, so an
-        element whose repr embeds an object address (or otherwise varies per
-        run) produces a key that never matches on reload. Prefer ints, floats,
-        bools, strs, enum ``.name``/``.value``, and tuples of those; avoid bare
-        objects, and note that ``file_key`` intentionally excludes
-        ``runner_hash``, so anything the persisted cache must distinguish has
-        to appear here rather than only in ``__hash__``.
+        Elements must also have a stable ``str()`` across processes because
+        ``ProfilingCacheKey.file_key`` serializes this tuple with ``str()``.
+        Prefer scalar values and enums over objects with address-dependent
+        representations. Since ``file_key`` excludes ``runner_hash``, persisted
+        distinctions must appear here rather than only in ``__hash__``.
         """
         return ()
